+++ modules/local/prepare_prediction_input/templates/prepare_prediction_input.py
@@ -26,6 +26,8 @@ class MinLength(Enum):
     MHCNUGGETS = 5
     NETMHCPAN = 8
     NETMHCIIPAN = 9
+    MIXMHCPRED = 8
+    MIXMHCIIPRED = 12
 
 class MaxLength(Enum):
     MHCFLURRY = 15
@@ -33,6 +35,8 @@ class MaxLength(Enum):
     MHCNUGGETS_CLASSII = 30
     NETMHCPAN = 14
     NETMHCIIPAN = 50
+    MIXMHCPRED = 14
+    MIXMHCIIPRED = 21
 
 # TODO: Implement
 class MaxNumberOfAlleles(Enum):
@@ -207,6 +211,8 @@ def main():
         "mhcnuggetsii": {"min": MinLength.MHCNUGGETS.value,  "max": MaxLength.MHCNUGGETS_CLASSII.value, "suffix": "mhcnuggetsii_input.tsv", "mhc_class": "II"},
         "netmhcpan":    {"min": MinLength.NETMHCPAN.value,   "max": MaxLength.NETMHCPAN.value,          "suffix": "netmhcpan_input.tsv",    "mhc_class": "I"},
         "netmhciipan":  {"min": MinLength.NETMHCIIPAN.value, "max": MaxLength.NETMHCIIPAN.value,        "suffix": "netmhciipan_input.tsv",  "mhc_class": "II"},
+        "mixmhcpred":   {"min": MinLength.MIXMHCPRED.value,  "max": MaxLength.MIXMHCPRED.value,         "suffix": "mixmhcpred_input.tsv",   "mhc_class": "I"},
+        "mixmhciipred": {"min": MinLength.MIXMHCIIPRED.value, "max": MaxLength.MIXMHCIIPRED.value,      "suffix": "mixmhciipred_input.tsv", "mhc_class": "II"},
     }
 
     # Step 2: Apply tool-specific length filtering** on top of MHC class filtering